Refactor helpblock, add tests, fix tests
Refactored mainhelp and travelhelp block into single block, updated all tests, fixed breaking ones
