+++ src/applications/check-in/day-of/tests/e2e/travel-path/travel.pay.path.cypress.spec.js
@@ -71,7 +71,6 @@ describe('Check In Experience', () => {
       cy.createScreenshots('Day-of-check-in--travel-pay--appointments');
       Appointments.attemptCheckIn(1);
       Confirmation.validatePageLoadedWithBtsssSubmission();
-      Confirmation.validateExtraFooterMessage();
       cy.injectAxeThenAxeCheck();
       cy.createScreenshots('Day-of-check-in--travel-pay--confirmation-success');
     });
